bit leaving works nice

+++ src/types.h
@@ -18,7 +18,7 @@ using flowfield_t = Eigen::ArrayXXd;
 using point_t = Eigen::Vector2d;
 using vector_t = Eigen::Vector2d;
 
-using handle_t = size_t; // not 0 and max_size
+using handle_t = uint64_t; // not 0 and max_size
 
 typedef struct coordinate {
     unsigned x;
